requirements.txt
absl-py==0.10.0
aiofiles==0.7.0
astor==0.8.1
attrs==20.2.0
bert4keras==0.10.6
boto==2.49.0
boto3==1.17.69
botocore==1.20.69
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1613247269562/work
chardet==3.0.4
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
contextvars==2.4
contrib==0.3.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
docker==5.0.0
english==2020.7.0
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
gast==0.2.2
gensim==3.8.1
gevent @ file:///C:/ci/gevent_1616773190302/work
gitdb==4.0.5
GitPython==3.1.11
google-pasta==0.2.0
greenlet @ file:///C:/ci/greenlet_1611958382984/work
grpcio==1.31.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.12.22
httptools==0.2.0
httpx==0.9.3
hyperframe==5.2.0
hyperopt==0.1.2
idna==2.10
immutables==0.15
importlib-metadata==1.7.0
isodate==0.6.0
itchat==1.3.10
itsdangerous @ file:///home/ktietz/src/ci/itsdangerous_1611932585308/work
jieba==0.39
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matchzoo-py==1.1.1
matplotlib==3.3.3
mock==4.0.2
monotonic==1.6
multidict==4.7.6
neo4j==4.1.1
neotime==1.7.4
networkx==2.5
numpy==1.17.3
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.1
pansi==2020.7.3
Pillow==8.0.1
prompt-toolkit==3.0.18
protobuf==3.13.0
py2neo==2021.1.1
pyahocorasick==1.4.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.9.0
pymongo==3.11.0
PyMySQL==0.10.1
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.25.1
rfc3986==1.5.0
s3transfer==0.4.2
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-openapi==0.6.2
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.1
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
smart-open==1.9.0
smmap==3.0.4
sniffio==1.2.0
SQLAlchemy==1.3.19
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
thulac==0.2.1
tqdm==4.44.1
ujson==1.35
urllib3==1.26.4
wcwidth==0.2.5
websocket-client==0.59.0
websockets==8.1
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
zmq==0.0.0
zope.event==4.5.0
zope.interface @ file:///C:/ci/zope.interface_1616357422405/work
